+++ src/components/space/spaceReusables/spaceCard/index.tsx
