--- src/components/space/spaceSidebar/spaceCard/SpaceCard.tsx
@@ -1,205 +1,0 @@
-// React + Web3 Essentials
-import React from 'react';
-
-// External Packages
-import styled, { useTheme } from 'styled-components';
-
-// Internal Components
-import { ImageV2, ItemHV2, ItemVV2, SpanV2 } from 'components/reusables/SharedStylingV2';
-import Calendar from 'assets/space/calendar.svg';
-import { shortenText } from 'helpers/UtilityHelper';
-import { Space } from 'types';
-import { Button } from 'primaries/SharedStyling';
-import { getSpaceTime, isSpaceActive } from 'helpers/space';
-
-interface SpaceCardType {
-  spaceData: Space;
-  borderRadius: string;
-  showActions?: boolean;
-}
-
-const SpaceCard = ({ spaceData, borderRadius, showActions }: SpaceCardType) => {
-  const activeSpace = isSpaceActive(spaceData?.scheduleAt, spaceData?.scheduleEnd);
-
-  const theme = useTheme();
-
-  const remindMe = () => {
-    console.log('Will remind');
-  };
-
-  const joinSpace = () => {
-    console.log('Joining');
-  };
-
-  return (
-    <SpaceCardContainer
-      borderRadius={borderRadius}
-      background={activeSpace ? theme.space.liveSpaceCardBackground : theme.space.scheduledSpaceCardBackground}
-    >
-      <SpaceCardHeader>
-        <ItemVV2
-          width="32px"
-          maxWidth="32px"
-          borderRadius="100%"
-          overflow="hidden"
-          margin="0 5px 0 0"
-        >
-          <ImageV2
-            alt="Profile"
-            src={spaceData?.spaceImage}
-          />
-        </ItemVV2>
-        <SpanV2
-          margin="0px 11px 0px 0px"
-          fontSize="17px"
-          fontWeight="500"
-          color="#fff"
-        >
-          {shortenText(spaceData?.spaceCreator, 6)}
-        </SpanV2>
-        <HostBadge>Host</HostBadge>
-      </SpaceCardHeader>
-      <SpanV2
-        fontSize="20px"
-        fontWeight="500"
-        color="#fff"
-        textAlign="left"
-        margin="24px 0px 0px 0px"
-      >
-        {!activeSpace
-          ? spaceData?.spaceName?.length > 45
-            ? spaceData?.spaceName?.slice(0, 43) + '...'
-            : spaceData?.spaceName
-          : `${shortenText(spaceData?.spaceCreator, 6)}'s Space`}
-        {/* {spaceData?.spaceName?.length > 45 ? spaceData?.spaceName?.slice(0, 43) + '...' : spaceData?.spaceName} */}
-      </SpanV2>
-      <SpaceCardFooter>
-        {activeSpace ? (
-          <LiveSpaceData>
-            <ItemHV2 justifyContent="flex-start">
-              <SpaceIcon />
-              <SpanV2
-                color="#F5F5F5"
-                fontSize="14px"
-                fontWeight="500"
-              >
-                Live
-              </SpanV2>
-            </ItemHV2>
-            <ItemHV2
-              justifyContent="flex-end"
-              alignItems="center"
-            >
-              {spaceData?.members?.slice(0, 3).map((member, index) => {
-                return (
-                  <ItemVV2
-                    width="31px"
-                    height="31px"
-                    maxWidth="31px"
-                    borderRadius="100%"
-                    overflow="hidden"
-                    margin="0px 0px 0px -18px"
-                    zIndex={4 - index}
-                    key={index}
-                  >
-                    <ImageV2
-                      height="100%"
-                      objectFit="cover"
-                      src={member?.image}
-                      alt="Member Image"
-                    />
-                  </ItemVV2>
-                );
-              })}
-              <SpanV2
-                color="#FFFFFF"
-                fontSize="14px"
-                fontWeight="500"
-                margin="0px 0px 0px 3px"
-              >
-                +{spaceData?.members?.length - 3}
-              </SpanV2>
-            </ItemHV2>
-          </LiveSpaceData>
-        ) : (
-          <ItemHV2 alignSelf="baseline">
-            <ImageV2
-              width="20px"
-              maxWidth="20px"
-              borderRadius="100%"
-              overflow="hidden"
-              margin="0 4px 0 0"
-              alt="Calendar"
-              src={Calendar}
-            />
-            <SpanV2
-              fontSize="14px"
-              fontWeight="500"
-              color="#fff"
-            >
-              {spaceData?.scheduleAt ? getSpaceTime(spaceData?.scheduleAt) : '30 Apr 4:30PM'}
-            </SpanV2>
-          </ItemHV2>
-        )}
-
-        {showActions && (
-          <ItemHV2 margin="10px 0 0 0">
-            <ActionButton
-              onClick={() => {
-                activeSpace ? joinSpace() : remindMe();
-              }}
-            >
-              {activeSpace ? 'Join this space' : 'Remind Me'}
-            </ActionButton>
-          </ItemHV2>
-        )}
-      </SpaceCardFooter>
-    </SpaceCardContainer>
-  );
-};
-
-const HostBadge = styled.div`
-  font-size: 12px;
-  font-weight: 500;
-  color: #fff;
-  padding: 4px 8px;
-  background: rgba(255, 255, 255, 0.2);
-  border-radius: 8px;
-`;
-
-const SpaceCardFooter = styled(ItemVV2)`
-  justify-content: flex-start;
-  // max-height: 20px;
-  margin-top: 12px;
-`;
-
-const SpaceCardHeader = styled(ItemHV2)`
-  justify-content: flex-start;
-  max-height: 32px;
-`;
-
-const SpaceCardContainer = styled(ItemVV2)`
-  width: 100%;
-  box-sizing: border-box;
-  justify-content: flex-start;
-  // max-height: 169px;
-  // background: linear-gradient(87.17deg, #b6a0f5 0%, #f46ef7 57.29%, #ff95d5 100%);
-  // border-radius: 17px;
-  padding: 16px;
-  // margin-bottom: 10px;
-`;
-
-const LiveSpaceData = styled(ItemHV2)`
-  justify-content: space-between;
-  max-height: 31px;
-  align-items: center;
-`;
-
-const ActionButton = styled(Button)`
-  border: 1px solid #ffffff;
-  border-radius: 8px;
-  background: transparent;
-  width: 100%;
-`;
-
-export default SpaceCard;
